Handle authentication state changes in main.js
This will handle things like page reload and login. This makes sure that
Firebase initializes before loading the app.

Note: made no-unused-vars lint rule a warning for now as the tutorial
adds a 'user' var but it's not used at this point. Might want to revert
this back to error always.