Login: Create the log-in logic.
Use the Firebase auth API to log in a user with email and password. This
is triggered by a click event on the login button in the page.

Set up Vuex with state data for the logged in user, including mutations
and actions. Fetch a user profile from the Firebase db (Firestore) users
collection we created previously and commit it to state.