Handle user state on page reload.
The local state is cleared when a page is reloaded, causing us to lose
user state data. Handle this through Firebase auth's onAuthStateChanged
by grabbing the available user data for logged in users and repopulating
it in the app.

NOTE:
In the tutorial (https://savvyapps.com/blog/definitive-guide-building-web-app-vuejs-firebase),
an onAuthStateChanged() observer was adedd in store/index.js but this
doesn't see to get called on reload. Probably because the tutorial is 2
years old and some aspects are a bit outdated.