Dashboard: Display user posts in the DOM.
Set up a query and snapshot for posts in store/index.js which will get
updated every time there are changes to the posts collection. Use
orderBy() so the posts come back newest first. Then loop through and add
the posts data to the state with a new mutation and state data.

Access the posts data in the dashboard through mapState, then add
conditional rendering (v-if/else) to show the posts.

Also made fixups to move onAuthStateChanged() operations that interact
with the store to store/index.js (things like re-populating the signed
in user data on page reload, or updating the posts from the query
snapshot)